[FIX] point_of_sale: image instead of image_medium
The image of the form view is set to be `image_medium`. However, the
method `image_resize_images` which is called during create or write
won't resize the image provided.

Therefore, the `image_medium` is not resized, leadnig to potential
issues if the user uploaded very large images.

We change the view to `image`, so that all images are properly resized.

opw-776943