[REV] base: revert "[IMP] base: no image (and kanban use default values"
This reverts commit c3332794a4b72f409cc11689433902c1f3417ff9.

It is not clear why it has been done. Original author thinks it can
be reverted as it makes Odoo have people without image which is not
good when starting with Odoo. We don't know what we want but what is
certain is that we need images instead of ugly void placeholder.