[FIX] stock: location created with wrong company.
When creating a new company the system will generate a
new warehouse, its location and its routes.
However the default company is the company that the user is
using.
For example user is in company_1 and creates company_2 it will create:
'company_2: Receipt in 1 step' and the company_id on it will be company_1

This commit completes the dict for create routes  in order to introduce
the key for company_id instead of using default value.