[FIX] mrp: default view of stock move line
This view was introduced by rev[1] to be more adequate in the usecase
of the finished move line in mrp. A priority was not set and thus it
became the default view for stock move line, even in the stock
application. Moreover, it used a field only used in mrp in a domain
resulting traceback when displayed in the context of stock.

[1] 665dde0e41a9fe1572cd5e589f5d378232ca2a51

Fixes #20610